@@ -5,17 +5,12 @@ import no.acntech.reservation.model.PendingReservationDto;
 import no.acntech.reservation.model.ReservationDto;
 import no.acntech.reservation.model.UpdateReservationDto;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
+import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -24,64 +19,75 @@ import java.util.UUID;
 @Component
 public class ReservationRestConsumer {
 
-    private static final ParameterizedTypeReference<List<ReservationDto>> RESERVATION_LIST = new ParameterizedTypeReference<List<ReservationDto>>() {
-    };
-    private final RestTemplate restTemplate;
+    private final WebClient webClient;
     private final String url;
 
-    public ReservationRestConsumer(final RestTemplateBuilder restTemplateBuilder,
-                                   @Value("${acntech.service.warehouse.api.reservations.url}") final String url) {
-        this.restTemplate = restTemplateBuilder.build();
+    public ReservationRestConsumer(final WebClient webClient,
+                                   @Value("${acntech.service.warehouse.url}/api/reservations") final String url) {
+        this.webClient = webClient;
         this.url = url;
     }
 
     public List<ReservationDto> find() {
-        final URI uri = UriComponentsBuilder.fromUriString(url)
+        final var uri = UriComponentsBuilder.fromUriString(url)
                 .build()
                 .toUri();
 
-        final ResponseEntity<List<ReservationDto>> entity = restTemplate.exchange(uri, HttpMethod.GET, null, RESERVATION_LIST);
-
-        return entity.getBody();
+        return webClient.get()
+                .uri(uri)
+                .retrieve()
+                .bodyToFlux(ReservationDto.class)
+                .collectList()
+                .block();
     }
 
     public Optional<ReservationDto> get(@NotNull final UUID reservationId) {
-        final URI uri = UriComponentsBuilder.fromUriString(url)
+        final var uri = UriComponentsBuilder.fromUriString(url)
                 .pathSegment(reservationId.toString())
                 .build()
                 .toUri();
 
-        final ResponseEntity<ReservationDto> entity = restTemplate.getForEntity(uri, ReservationDto.class);
-        return Optional.of(entity)
-                .map(ResponseEntity::getBody);
+        return webClient.get()
+                .uri(uri)
+                .retrieve()
+                .bodyToMono(ReservationDto.class)
+                .blockOptional();
     }
 
-    public Optional<PendingReservationDto> create(@Valid final CreateReservationDto createReservation) {
-        final URI uri = UriComponentsBuilder.fromUriString(url)
+    public Optional<PendingReservationDto> create(@Valid final CreateReservationDto createReservationDto) {
+        final var uri = UriComponentsBuilder.fromUriString(url)
                 .build()
                 .toUri();
 
-        ResponseEntity<PendingReservationDto> entity = restTemplate.postForEntity(uri, createReservation, PendingReservationDto.class);
-        return Optional.of(entity)
-                .map(ResponseEntity::getBody);
+        return webClient.post()
+                .uri(uri)
+                .bodyValue(createReservationDto)
+                .retrieve()
+                .bodyToMono(PendingReservationDto.class)
+                .blockOptional();
     }
 
     public void update(@NotNull final UUID reservationId,
-                       @Valid final UpdateReservationDto updateReservation) {
-        final URI uri = UriComponentsBuilder.fromUriString(url)
+                       @Valid final UpdateReservationDto updateReservationDto) {
+        final var uri = UriComponentsBuilder.fromUriString(url)
                 .pathSegment(reservationId.toString())
                 .build()
                 .toUri();
 
-        restTemplate.put(uri, updateReservation);
+        webClient.put()
+                .uri(uri)
+                .bodyValue(updateReservationDto)
+                .retrieve();
     }
 
     public void delete(@NotNull final UUID reservationId) {
-        final URI uri = UriComponentsBuilder.fromUriString(url)
+        final var uri = UriComponentsBuilder.fromUriString(url)
                 .pathSegment(reservationId.toString())
                 .build()
                 .toUri();
 
-        restTemplate.delete(uri);
+        webClient.delete()
+                .uri(uri)
+                .retrieve();
     }
 }
