@@ -1,30 +1,26 @@
 package no.acntech.order.consumer;
 
-import javax.validation.constraints.NotNull;
+import no.acntech.order.model.OrderDto;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.util.UriComponentsBuilder;
 
+import javax.validation.constraints.NotNull;
 import java.net.URI;
 import java.util.Optional;
 import java.util.UUID;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import no.acntech.order.model.OrderDto;
 
 @SuppressWarnings("Duplicates")
 @Component
 public class OrderRestConsumer {
 
-    private final RestTemplate restTemplate;
+    private final WebClient webClient;
     private final String url;
 
-    public OrderRestConsumer(final RestTemplateBuilder restTemplateBuilder,
-                             @Value("${acntech.service.ordering.api.orders.url}") final String url) {
-        this.restTemplate = restTemplateBuilder.build();
+    public OrderRestConsumer(final WebClient webClient,
+                             @Value("${acntech.service.ordering.url}/api/orders") final String url) {
+        this.webClient = webClient;
         this.url = url;
     }
 
@@ -34,8 +30,10 @@ public class OrderRestConsumer {
                 .build()
                 .toUri();
 
-        final ResponseEntity<OrderDto> entity = restTemplate.getForEntity(uri, OrderDto.class);
-        return Optional.of(entity)
-                .map(ResponseEntity::getBody);
+        return webClient.get()
+                .uri(uri)
+                .retrieve()
+                .bodyToMono(OrderDto.class)
+                .blockOptional();
     }
 }
