@@ -1,31 +1,24 @@
 package no.acntech.common.config;
 
-import java.util.Arrays;
-
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import no.acntech.order.model.OrderEvent;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-
-import no.acntech.order.model.OrderEvent;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.config.KafkaListenerContainerFactory;
+import org.springframework.kafka.core.*;
+import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
 
 @SuppressWarnings("Duplicates")
 @EnableKafka
 @Configuration
 public class KafkaConfig {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(KafkaConfig.class);
     private final KafkaProperties kafkaProperties;
 
     public KafkaConfig(final KafkaProperties kafkaProperties) {
         this.kafkaProperties = kafkaProperties;
-        LOGGER.info("Configuring Kafka, connecting to bootstrap servers {}...", Arrays.toString(kafkaProperties.getBootstrapServers().toArray()));
     }
 
     @Bean
@@ -39,7 +32,14 @@ public class KafkaConfig {
     }
 
     @Bean
-    public KafkaConsumer<String, OrderEvent> orderKafkaConsumer() {
-        return new KafkaConsumer<>(kafkaProperties.buildConsumerProperties());
+    public ConsumerFactory<String, OrderEvent> ordersKafkaConsumerFactory() {
+        return new DefaultKafkaConsumerFactory<>(kafkaProperties.buildConsumerProperties());
+    }
+
+    @Bean
+    public KafkaListenerContainerFactory<ConcurrentMessageListenerContainer<String, OrderEvent>> ordersKafkaListenerContainerFactory() {
+        final var factory = new ConcurrentKafkaListenerContainerFactory<String, OrderEvent>();
+        factory.setConsumerFactory(ordersKafkaConsumerFactory());
+        return factory;
     }
 }
