@@ -1,22 +1,25 @@
 package no.acntech.invoice.service;
 
+import no.acntech.invoice.exception.InvoiceNotFoundException;
+import no.acntech.invoice.model.CreateInvoiceDto;
+import no.acntech.invoice.model.InvoiceDto;
+import no.acntech.invoice.model.InvoiceEntity;
+import no.acntech.invoice.model.InvoiceQuery;
+import no.acntech.invoice.repository.InvoiceRepository;
+import org.springframework.core.convert.ConversionService;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.Assert;
+import org.springframework.validation.annotation.Validated;
+
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import org.springframework.core.convert.ConversionService;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
-
-import no.acntech.invoice.exception.InvoiceNotFoundException;
-import no.acntech.invoice.model.CreateInvoiceDto;
-import no.acntech.invoice.model.Invoice;
-import no.acntech.invoice.model.InvoiceDto;
-import no.acntech.invoice.model.InvoiceQuery;
-import no.acntech.invoice.model.InvoiceStatus;
-import no.acntech.invoice.repository.InvoiceRepository;
-
+@Validated
 @Service
 public class InvoiceService {
 
@@ -31,21 +34,19 @@ public class InvoiceService {
     }
 
     @SuppressWarnings("Duplicates")
-    public List<InvoiceDto> findInvoices(final InvoiceQuery invoiceQuery) {
-        UUID orderId = invoiceQuery.getOrderId();
-        InvoiceStatus status = invoiceQuery.getStatus();
-        if (orderId != null && status != null) {
-            return invoiceRepository.findAllByOrderIdAndStatus(orderId, status)
+    public List<InvoiceDto> findInvoices(@NotNull @Valid final InvoiceQuery invoiceQuery) {
+        if (invoiceQuery.getOrderId() != null && invoiceQuery.getStatus() != null) {
+            return invoiceRepository.findAllByOrderIdAndStatus(invoiceQuery.getOrderId(), invoiceQuery.getStatus())
                     .stream()
                     .map(this::convert)
                     .collect(Collectors.toList());
-        } else if (orderId != null) {
-            return invoiceRepository.findAllByOrderId(orderId)
+        } else if (invoiceQuery.getOrderId() != null) {
+            return invoiceRepository.findAllByOrderId(invoiceQuery.getOrderId())
                     .stream()
                     .map(this::convert)
                     .collect(Collectors.toList());
-        } else if (status != null) {
-            return invoiceRepository.findAllByStatus(status)
+        } else if (invoiceQuery.getStatus() != null) {
+            return invoiceRepository.findAllByStatus(invoiceQuery.getStatus())
                     .stream()
                     .map(this::convert)
                     .collect(Collectors.toList());
@@ -57,22 +58,25 @@ public class InvoiceService {
         }
     }
 
-    public InvoiceDto getInvoice(final UUID invoiceId) {
+    public InvoiceDto getInvoice(@NotNull final UUID invoiceId) {
         return invoiceRepository.findByInvoiceId(invoiceId)
                 .map(this::convert)
                 .orElseThrow(() -> new InvoiceNotFoundException(invoiceId));
     }
 
-    public InvoiceDto createInvoice(final CreateInvoiceDto createInvoice) {
-        Invoice invoice = conversionService.convert(createInvoice, Invoice.class);
-        Assert.notNull(invoice, "Failed to convert invoice");
-
-        Invoice savedInvoice = invoiceRepository.save(invoice);
+    @SuppressWarnings("UnusedReturnValue")
+    @Transactional
+    public InvoiceDto createInvoice(@NotNull @Valid final CreateInvoiceDto createInvoiceDto) {
+        final var invoiceEntity = conversionService.convert(createInvoiceDto, InvoiceEntity.class);
+        Assert.notNull(invoiceEntity, "Failed to convert CreateInvoiceDto to InvoiceEntity");
+        final var savedInvoice = invoiceRepository.save(invoiceEntity);
         return convert(savedInvoice);
     }
 
-    private InvoiceDto convert(final Invoice invoice) {
-        return conversionService.convert(invoice, InvoiceDto.class);
+    private InvoiceDto convert(final InvoiceEntity invoiceEntity) {
+        final var invoiceDto = conversionService.convert(invoiceEntity, InvoiceDto.class);
+        Assert.notNull(invoiceDto, "Failed to convert InvoiceEntity to InvoiceDto");
+        return invoiceDto;
     }
 }
 
