@@ -20,7 +20,7 @@ public class OrderDto {
     @NotNull
     private OrderStatus status;
     @NotNull
-    private List<ItemDto> items;
+    private List<OrderItemDto> items;
     @NotNull
     private ZonedDateTime modified;
     private ZonedDateTime created;
@@ -45,7 +45,7 @@ public class OrderDto {
         return status;
     }
 
-    public List<ItemDto> getItems() {
+    public List<OrderItemDto> getItems() {
         return items;
     }
 
