@@ -1,70 +1,49 @@
 package no.acntech.customer.resource;
 
-import javax.validation.Valid;
-
-import java.net.URI;
-import java.util.List;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
-import org.springframework.core.convert.ConversionService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 import no.acntech.customer.model.CreateCustomerDto;
-import no.acntech.customer.model.Customer;
 import no.acntech.customer.model.CustomerDto;
 import no.acntech.customer.model.CustomerQuery;
 import no.acntech.customer.service.CustomerService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-@RequestMapping(path = "customers")
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping(path = "/api/customers")
 @RestController
 public class CustomersResource {
 
-    private final ConversionService conversionService;
     private final CustomerService customerService;
 
-    public CustomersResource(final ConversionService conversionService,
-                             final CustomerService customerService) {
-        this.conversionService = conversionService;
+    public CustomersResource(final CustomerService customerService) {
         this.customerService = customerService;
     }
 
     @GetMapping("{customerId}")
     public ResponseEntity<CustomerDto> get(@PathVariable("customerId") final UUID customerId) {
-        final Customer customer = customerService.getCustomer(customerId);
-        final CustomerDto customerDto = convert(customer);
+        final var customerDto = customerService.getCustomer(customerId);
         return ResponseEntity.ok(customerDto);
 
     }
 
     @GetMapping
     public ResponseEntity<List<CustomerDto>> find(final CustomerQuery customerQuery) {
-        final List<Customer> customers = customerService.findCustomers(customerQuery);
-        final List<CustomerDto> customerDtos = customers.stream()
-                .map(this::convert)
-                .collect(Collectors.toList());
+        final var customerDtos = customerService.findCustomers(customerQuery);
         return ResponseEntity.ok(customerDtos);
     }
 
     @PostMapping
-    public ResponseEntity post(@Valid @RequestBody final CreateCustomerDto createCustomer) {
-        final Customer customer = customerService.createCustomer(createCustomer);
-        URI location = ServletUriComponentsBuilder
+    public ResponseEntity<CustomerDto> post(@RequestBody final CreateCustomerDto createCustomerDto) {
+        final var customerDto = customerService.createCustomer(createCustomerDto);
+        final var location = ServletUriComponentsBuilder
                 .fromCurrentRequest()
-                .path("/{customerId}")
-                .buildAndExpand(customer.getCustomerId())
+                .pathSegment(customerDto.getCustomerId().toString())
+                .build()
                 .toUri();
-        return ResponseEntity.created(location).build();
-    }
-
-    private CustomerDto convert(final Customer customer) {
-        return conversionService.convert(customer, CustomerDto.class);
+        return ResponseEntity
+                .created(location)
+                .body(customerDto);
     }
 }
