@@ -2,22 +2,13 @@ package no.acntech;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-@EnableAsync
 @EnableTransactionManagement
 @SpringBootApplication
-public class EventNotificationOrderingApplication extends SpringBootServletInitializer {
+public class EventNotificationOrderingApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(EventNotificationOrderingApplication.class, args);
-    }
-
-    @Override
-    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-        return application.sources(EventNotificationOrderingApplication.class);
     }
 }
