@@ -1,45 +1,51 @@
 package no.acntech.common.config;
 
-import java.util.Arrays;
-
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import no.acntech.reservation.model.ReservationEvent;
+import org.apache.kafka.clients.admin.NewTopic;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-
-import no.acntech.reservation.model.ReservationEvent;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.config.KafkaListenerContainerFactory;
+import org.springframework.kafka.core.*;
+import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
 
 @SuppressWarnings("Duplicates")
 @EnableKafka
 @Configuration
 public class KafkaConfig {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(KafkaConfig.class);
     private final KafkaProperties kafkaProperties;
 
     public KafkaConfig(final KafkaProperties kafkaProperties) {
         this.kafkaProperties = kafkaProperties;
-        LOGGER.info("Configuring Kafka, connecting to bootstrap servers {}...", Arrays.toString(kafkaProperties.getBootstrapServers().toArray()));
+    }
+
+    @Bean
+    public NewTopic reservationsTopic() {
+        return new NewTopic(KafkaTopic.RESERVATIONS, 1, (short) 1);
     }
 
     @Bean
     public KafkaTemplate<String, Object> kafkaTemplate() {
-        return new KafkaTemplate<>(producerFactory());
+        return new KafkaTemplate<>(kafkaProducerFactory());
     }
 
     @Bean
-    public ProducerFactory<String, Object> producerFactory() {
+    public ProducerFactory<String, Object> kafkaProducerFactory() {
         return new DefaultKafkaProducerFactory<>(kafkaProperties.buildProducerProperties());
     }
 
     @Bean
-    public KafkaConsumer<String, ReservationEvent> reservationKafkaConsumer() {
-        return new KafkaConsumer<>(kafkaProperties.buildConsumerProperties());
+    public ConsumerFactory<String, ReservationEvent> reservationKafkaConsumerFactory() {
+        return new DefaultKafkaConsumerFactory<>(kafkaProperties.buildConsumerProperties());
+    }
+
+    @Bean
+    public KafkaListenerContainerFactory<ConcurrentMessageListenerContainer<String, ReservationEvent>> reservationKafkaListenerContainerFactory() {
+        final var factory = new ConcurrentKafkaListenerContainerFactory<String, ReservationEvent>();
+        factory.setConsumerFactory(reservationKafkaConsumerFactory());
+        return factory;
     }
 }
