@@ -1,12 +1,9 @@
 package no.acntech.order.model;
 
-import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-
 import java.io.Serializable;
 import java.util.UUID;
 
-@Valid
 public class OrderEvent implements Serializable {
 
     @NotNull
@@ -33,9 +30,9 @@ public class OrderEvent implements Serializable {
         }
 
         public OrderEvent build() {
-            OrderEvent orderEvent = new OrderEvent();
-            orderEvent.orderId = this.orderId;
-            return orderEvent;
+            final var target = new OrderEvent();
+            target.orderId = this.orderId;
+            return target;
         }
     }
 }
