@@ -1,23 +1,20 @@
 package no.acntech.order.producer;
 
-import javax.validation.constraints.NotNull;
-
-import java.util.UUID;
-
+import no.acntech.common.config.KafkaTopic;
+import no.acntech.order.model.OrderEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import no.acntech.common.config.KafkaTopic;
-import no.acntech.order.model.OrderEvent;
+import javax.validation.constraints.NotNull;
+import java.util.UUID;
 
 @Component
 public class OrderEventProducer {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(OrderEventProducer.class);
-
     private final KafkaTemplate<String, Object> kafkaTemplate;
 
     public OrderEventProducer(final KafkaTemplate<String, Object> kafkaTemplate) {
@@ -26,11 +23,11 @@ public class OrderEventProducer {
 
     @Transactional
     public void publish(@NotNull final UUID orderId) {
-        LOGGER.debug("Sending event for order-id {} to topic {}", orderId, KafkaTopic.ORDERS.getName());
+        LOGGER.debug("Sending order event for order-id {} to topic {}", orderId, KafkaTopic.ORDERS);
         OrderEvent orderEvent = OrderEvent.builder()
                 .orderId(orderId)
                 .build();
-        kafkaTemplate.send(KafkaTopic.ORDERS.getName(), orderEvent);
+        kafkaTemplate.send(KafkaTopic.ORDERS, orderEvent);
     }
 }
 
