@@ -2,7 +2,6 @@ package no.acntech.order.model;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -20,7 +19,7 @@ public class OrderDto {
     @NotNull
     private OrderStatus status;
     @NotNull
-    private List<ItemDto> items;
+    private List<OrderItemDto> items;
     @NotNull
     private ZonedDateTime modified;
     private ZonedDateTime created;
@@ -45,7 +44,7 @@ public class OrderDto {
         return status;
     }
 
-    public List<ItemDto> getItems() {
+    public List<OrderItemDto> getItems() {
         return items;
     }
 
