@@ -7,23 +7,23 @@ import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import no.acntech.shipment.model.Shipment;
+import no.acntech.shipment.model.ShipmentEntity;
 import no.acntech.shipment.model.ShipmentStatus;
 
 @Repository
-public interface ShipmentRepository extends JpaRepository<Shipment, Long> {
+public interface ShipmentRepository extends JpaRepository<ShipmentEntity, Long> {
 
-    Optional<Shipment> findByShipmentId(UUID shipmentId);
+    Optional<ShipmentEntity> findByShipmentId(UUID shipmentId);
 
-    List<Shipment> findAllByCustomerId(UUID customerId);
+    List<ShipmentEntity> findAllByCustomerId(UUID customerId);
 
-    List<Shipment> findAllByOrderId(UUID orderId);
+    List<ShipmentEntity> findAllByOrderId(UUID orderId);
 
-    List<Shipment> findAllByStatus(ShipmentStatus status);
+    List<ShipmentEntity> findAllByStatus(ShipmentStatus status);
 
-    List<Shipment> findAllByCustomerIdAndStatus(UUID customerId, ShipmentStatus status);
+    List<ShipmentEntity> findAllByCustomerIdAndStatus(UUID customerId, ShipmentStatus status);
 
-    List<Shipment> findAllByOrderIdAndStatus(UUID orderId, ShipmentStatus status);
+    List<ShipmentEntity> findAllByOrderIdAndStatus(UUID orderId, ShipmentStatus status);
 
-    List<Shipment> findAllByCustomerIdAndOrderIdAndStatus(UUID customerId, UUID orderId, ShipmentStatus status);
+    List<ShipmentEntity> findAllByCustomerIdAndOrderIdAndStatus(UUID customerId, UUID orderId, ShipmentStatus status);
 }
