@@ -1,16 +1,11 @@
 package no.acntech.order.service;
 
-import java.util.Optional;
-import java.util.UUID;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Service;
-
 import no.acntech.order.consumer.OrderRestConsumer;
 import no.acntech.order.model.OrderDto;
 import no.acntech.order.model.OrderEvent;
-import no.acntech.order.model.OrderStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
 
 @Service
 public class OrderService {
@@ -24,31 +19,23 @@ public class OrderService {
 
     @SuppressWarnings("Duplicates")
     public void receiveOrderEvent(final OrderEvent orderEvent) {
-        UUID orderId = orderEvent.getOrderId();
-
-        LOGGER.debug("Fetching order for order-id {}", orderId);
+        LOGGER.debug("Fetching order for order-id {}", orderEvent.getOrderId());
 
         try {
-            Optional<OrderDto> orderOptional = orderRestConsumer.get(orderId);
-
-            if (orderOptional.isPresent()) {
-                OrderDto order = orderOptional.get();
-
-                processOrder(order);
+            final var optionalOrderDto = orderRestConsumer.get(orderEvent.getOrderId());
+            if (optionalOrderDto.isPresent()) {
+                final var orderDto = optionalOrderDto.get();
+                processOrder(orderDto);
             } else {
-                LOGGER.error("Order with order-id {} could not be found", orderId);
+                LOGGER.error("Order with order-id {} could not be found", orderEvent.getOrderId());
             }
         } catch (Exception e) {
             LOGGER.error("Error occurred while processing order", e);
         }
     }
 
-    private void processOrder(final OrderDto order) {
-        UUID orderId = order.getOrderId();
-        OrderStatus status = order.getStatus();
-
-        LOGGER.debug("Processing order for order-id {}", orderId);
-
-        LOGGER.debug("Ignoring order for order-id {} and status {}", orderId, status);
+    private void processOrder(final OrderDto orderDto) {
+        LOGGER.debug("Processing orderDto for orderDto-id {}", orderDto.getOrderId());
+        LOGGER.debug("Ignoring orderDto for orderDto-id {} and status {}", orderDto.getOrderId(), orderDto.getStatus());
     }
 }
