@@ -1,16 +1,14 @@
 package no.acntech.reservation.producer;
 
-import javax.validation.constraints.NotNull;
-
-import java.util.UUID;
-
+import no.acntech.common.config.KafkaTopic;
+import no.acntech.reservation.model.ReservationEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
-import no.acntech.common.config.KafkaTopic;
-import no.acntech.reservation.model.ReservationEvent;
+import javax.validation.constraints.NotNull;
+import java.util.UUID;
 
 @Component
 public class ReservationEventProducer {
@@ -24,10 +22,10 @@ public class ReservationEventProducer {
     }
 
     public void publish(@NotNull final UUID reservationId) {
-        LOGGER.debug("Sending event for reservation-id {} to topic {}", reservationId, KafkaTopic.RESERVATIONS.getName());
-        ReservationEvent reservationEvent = ReservationEvent.builder()
+        LOGGER.debug("Sending event for reservation-id {} to topic {}", reservationId, KafkaTopic.RESERVATIONS);
+        final var reservationEvent = ReservationEvent.builder()
                 .reservationId(reservationId)
                 .build();
-        kafkaTemplate.send(KafkaTopic.RESERVATIONS.getName(), reservationEvent);
+        kafkaTemplate.send(KafkaTopic.RESERVATIONS, reservationEvent);
     }
 }
