@@ -4,17 +4,16 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import no.acntech.reservation.model.ReservationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import no.acntech.reservation.model.Reservation;
-
 @Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<ReservationEntity, Long> {
 
-    Optional<Reservation> findByReservationId(UUID reservationId);
+    Optional<ReservationEntity> findByReservationId(UUID reservationId);
 
-    List<Reservation> findAllByOrderId(UUID orderId);
+    List<ReservationEntity> findAllByOrderId(UUID orderId);
 
-    Optional<Reservation> findByOrderIdAndProduct_ProductId(UUID orderId, UUID productId);
+    Optional<ReservationEntity> findByOrderIdAndProduct_ProductId(UUID orderId, UUID productId);
 }
