@@ -1,31 +1,24 @@
 package no.acntech.reservation.service;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
+import no.acntech.product.repository.ProductRepository;
+import no.acntech.reservation.exception.ReservationNotFoundException;
+import no.acntech.reservation.model.*;
+import no.acntech.reservation.repository.ReservationRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
 
-import no.acntech.product.model.Product;
-import no.acntech.product.repository.ProductRepository;
-import no.acntech.reservation.exception.ReservationNotFoundException;
-import no.acntech.reservation.model.CreateReservationDto;
-import no.acntech.reservation.model.PendingReservationDto;
-import no.acntech.reservation.model.Reservation;
-import no.acntech.reservation.model.ReservationDto;
-import no.acntech.reservation.model.ReservationStatus;
-import no.acntech.reservation.model.UpdateReservationDto;
-import no.acntech.reservation.repository.ReservationRepository;
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
 @SuppressWarnings({"Duplicates", "WeakerAccess"})
+@Validated
 @Service
 public class ReservationService {
 
@@ -63,62 +56,61 @@ public class ReservationService {
     }
 
     @Transactional
-    public void createReservation(@Valid final PendingReservationDto pendingReservation,
+    public void createReservation(@NotNull final UUID reservationId,
                                   @Valid final CreateReservationDto createReservation) {
-        UUID reservationId = pendingReservation.getReservationId();
-        UUID orderId = createReservation.getOrderId();
-        UUID productId = createReservation.getProductId();
-        Long quantity = createReservation.getQuantity();
+        final var orderId = createReservation.getOrderId();
+        final var productId = createReservation.getProductId();
+        final var quantity = createReservation.getQuantity();
 
-        Optional<Reservation> existingReservation = reservationRepository.findByOrderIdAndProduct_ProductId(orderId, productId);
+        final var optionalReservationEntity = reservationRepository
+                .findByOrderIdAndProduct_ProductId(orderId, productId);
 
-        if (existingReservation.isPresent()) {
-            Reservation reservation = Reservation.builder()
+        if (optionalReservationEntity.isPresent()) {
+            ReservationEntity reservation = ReservationEntity.builder()
                     .reservationId(reservationId)
                     .orderId(orderId)
                     .quantity(quantity)
                     .statusFailed()
                     .build();
             reservationRepository.save(reservation);
-
             LOGGER.error("Reservation already exists for order-id {} and product-id {}", orderId, productId);
         } else {
-            final Optional<Product> existingProduct = productRepository.findByProductId(productId);
+            final var optionalProductEntity = productRepository.findByProductId(productId);
 
-            if (existingProduct.isPresent()) {
-                Product product = existingProduct.get();
+            if (optionalProductEntity.isPresent()) {
+                final var productEntity = optionalProductEntity.get();
 
-                if (product.getStock() < quantity) {
-                    Reservation reservation = Reservation.builder()
+                if (productEntity.getStock() < quantity) {
+                    final var reservationEntity = ReservationEntity.builder()
                             .reservationId(reservationId)
                             .orderId(orderId)
-                            .product(product)
                             .quantity(quantity)
+                            .product(productEntity)
                             .statusRejected()
                             .build();
-                    reservationRepository.save(reservation);
+                    reservationRepository.save(reservationEntity);
 
                     LOGGER.error("Product stock insufficient for reservation-id {}", reservationId);
                 } else {
-                    Reservation reservation = Reservation.builder()
+                    final var reservationEntity = ReservationEntity.builder()
                             .reservationId(reservationId)
                             .orderId(orderId)
-                            .product(product)
                             .quantity(quantity)
+                            .product(productEntity)
                             .statusReserved()
                             .build();
-                    reservationRepository.save(reservation);
+                    reservationRepository.save(reservationEntity);
 
                     LOGGER.info("Created reservation for reservation-id {}", reservationId);
                 }
             } else {
-                Reservation reservation = Reservation.builder()
+                final var reservationEntity = ReservationEntity.builder()
                         .reservationId(reservationId)
                         .orderId(orderId)
                         .quantity(quantity)
                         .statusFailed()
                         .build();
-                reservationRepository.save(reservation);
+                reservationRepository.save(reservationEntity);
 
                 LOGGER.error("No product found for reservation-id {}", reservationId);
             }
@@ -128,50 +120,33 @@ public class ReservationService {
     @Transactional
     public void updateReservation(@NotNull final UUID reservationId,
                                   @Valid final UpdateReservationDto updateReservation) {
-        Long quantity = updateReservation.getQuantity();
-        ReservationStatus status = updateReservation.getStatus();
+        final var reservationEntity = reservationRepository.findByReservationId(reservationId)
+                .orElseThrow(() -> new ReservationNotFoundException(reservationId));
 
-        Optional<Reservation> existingReservation = reservationRepository.findByReservationId(reservationId);
-
-        if (existingReservation.isPresent()) {
-            Reservation reservation = existingReservation.get();
-
-            if (quantity != null) {
-                reservation.setQuantity(quantity);
-            }
-            if (status != null && canUpdateStatus(reservation.getStatus())) {
-                reservation.setStatus(status);
-            }
-
-            reservationRepository.save(reservation);
-
-            LOGGER.info("Updated reservation for reservation-id {}", reservationId);
-        } else {
-            throw new ReservationNotFoundException(reservationId);
+        if (updateReservation.getQuantity() != null) {
+            reservationEntity.setQuantity(updateReservation.getQuantity());
+        }
+        if (updateReservation.getStatus() != null && reservationEntity.canUpdateStatus()) {
+            reservationEntity.setStatus(updateReservation.getStatus());
         }
 
+        reservationRepository.save(reservationEntity);
+
+        LOGGER.info("Updated reservation for reservation-id {}", reservationId);
     }
 
     @Transactional
     public void deleteReservation(@NotNull final UUID reservationId) {
-        Optional<Reservation> existingReservation = reservationRepository.findByReservationId(reservationId);
+        final var reservationEntity = reservationRepository.findByReservationId(reservationId)
+                .orElseThrow(() -> new ReservationNotFoundException(reservationId));
 
-        if (existingReservation.isPresent()) {
-            Reservation reservation = existingReservation.get();
-            reservation.setStatus(ReservationStatus.CANCELED);
-            reservationRepository.save(reservation);
+        reservationEntity.setStatus(ReservationStatus.CANCELED);
+        reservationRepository.save(reservationEntity);
 
-            LOGGER.info("Updated reservation for reservation-id {}", reservationId);
-        } else {
-            throw new ReservationNotFoundException(reservationId);
-        }
+        LOGGER.info("Updated reservation for reservation-id {}", reservationId);
     }
 
-    private boolean canUpdateStatus(ReservationStatus status) {
-        return status != null && status != ReservationStatus.CANCELED && status != ReservationStatus.CONFIRMED;
-    }
-
-    private ReservationDto convert(final Reservation reservation) {
+    private ReservationDto convert(final ReservationEntity reservation) {
         return conversionService.convert(reservation, ReservationDto.class);
     }
 }
